NEW: Split KafkaST and do some refactor
